requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
xxhash>=3.4.0
datasketch>=1.6.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0
jsonlines>=4.0.0
tqdm>=4.66.0
pandas>=2.1.0
